pyproject.toml
[project]
name = "train-csie-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "python-dotenv>=1.0.1",
]
